requirements.txt
streamlit>=1.30.0
langgraph>=0.2.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
